ethtool: Remove end-to-end testing.
