+++ end-to-end-test.sh
@@ -14,7 +14,6 @@ enabled_collectors=$(cat << COLLECTORS
   drbd
   edac
   entropy
-  ethtool
   fibrechannel
   filefd
   hwmon
@@ -110,7 +109,6 @@ fi
   --collector.qdisc.fixtures="collector/fixtures/qdisc/" \
   --collector.netclass.ignored-devices="(dmz|int)" \
   --collector.netclass.ignore-invalid-speed \
-  --collector.ethtool.fixtures="collector/fixtures/ethtool/" \
   --collector.bcache.priorityStats \
   --collector.cpu.info \
   --collector.cpu.info.flags-include="^(aes|avx.?|constant_tsc)$" \
